# specs/001-create-a-rest/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Test Results Management API
  description: |
    REST API for CRUD operations on end-to-end test execution results.
    Primary integration via GitHub Actions CI/CD pipelines using bearer token authentication.

    ## Authentication
    All endpoints require JWT bearer token authentication. Tokens are obtained through
    GitHub SSO for human users or generated for automated systems (CI/CD pipelines).

    ## Integration Patterns
    - **GitHub Actions**: Use bearer tokens stored in repository/organization secrets
    - **CI/CD Systems**: Automated test result submission with artifact upload
    - **Development**: Local testing with personal access tokens
  version: 0.4.0
  contact:
    name: Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.testresults.dev
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TestResult:
      type: object
      required: [suite_id, test_name, full_title, status, duration_ms]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        suite_id:
          type: string
          format: uuid
        test_name:
          type: string
          maxLength: 500
          example: "manages focus when opening and closing settings modal with keyboard"
        full_title:
          type: string
          maxLength: 1000
          example: "Accessibility Support > Settings Dialog > manages focus when opening and closing settings modal with keyboard"
        status:
          type: string
          enum: [passed, failed, skipped, pending]
          example: "passed"
        duration_ms:
          type: integer
          minimum: 0
          example: 3971
        error_message:
          type: string
          nullable: true
          example: "Expected element to be visible"
        stack_trace:
          type: string
          nullable: true
          example: "Error: Expected element to be visible\\n    at test/settings.spec.ts:25:5"
        retry_count:
          type: integer
          minimum: 0
          default: 0
          example: 0
        tags:
          type: array
          items:
            type: string
          maxItems: 20
          example: ["accessibility", "settings"]
        external_id:
          type: string
          nullable: true
          example: "c27dc2d4fd7579bfae82-84088ebe5b793a8b0db8"
        metadata:
          type: object
          additionalProperties: true
          example:
            projectId: "ipad"
            projectName: "ipad"
            workerIndex: 1
            parallelIndex: 0
            file: "accessibility/channels/settings_dialog/settings.spec.ts"
            line: 9
            column: 5
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    PaginatedTestResults:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TestResult'
        total:
          type: integer
          example: 150
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        pages:
          type: integer
          example: 8
...
paths:
...
  /results:
    get:
      summary: List test results
      tags: [Results]
      parameters:
        - name: suite_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by test suite
        - name: test_name
          in: query
          schema:
            type: string
          description: Filter by test name (partial match)
        - name: status
          in: query
          schema:
            type: string
            enum: [passed, failed, skipped]
          description: Filter by test status
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
          description: Filter results created after this date
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
          description: Filter results created before this date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of test results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTestResults'

    post:
      summary: Create test result
      tags: [Results]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestResult'
      responses:
        '201':
          description: Result created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'
        '400':
          description: Invalid input
  /results/bulk:
    post:
      summary: Create multiple test results
      tags: [Results]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/TestResult'
              maxItems: 1000
      responses:
        '201':
          description: Results created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  created_count:
                    type: integer
                    example: 25
                  failed_count:
                    type: integer
                    example: 2
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        index:
                          type: integer
                        error:
                          type: string
        '400':
          description: Invalid input
  /results/{result_id}:
    get:
      summary: Get test result
      tags: [Results]
      parameters:
        - name: result_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Result details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'
        '404':
          description: Result not found

    put:
      summary: Update test result
      tags: [Results]
      parameters:
        - name: result_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestResult'
      responses:
        '200':
          description: Result updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResult'

    delete:
      summary: Delete test result
      tags: [Results]
      parameters:
        - name: result_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Result deleted successfully
        '404':
          description: Result not found

  # Authentication
...